# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Payment Gateway API
  version: "1.1.0"
  description: >
    USDT payment gateway supporting BEP20 (BNB Smart Chain) and TRC20 (Tron).
    Merchant endpoints authenticate with API key + HMAC signature; dashboard
    endpoints authenticate with JWT. The target chain is chosen per-payment via
    the optional `network` field (defaults to BEP20 for backward compatibility);
    TRC20 must be enabled server-side. Amounts are always human USDT strings —
    the gateway handles each chain's decimals (BEP20=18, TRC20=6) internally.
servers:
  - url: https://gateway.example.com/api/v1
  - url: http://localhost:4000/api/v1
...
components:
  securitySchemes:
    # Two API key modes coexist. Which one applies is decided by the key itself
    # (api_keys.auth_mode), not by the caller — presenting an HMAC key without a
    # signature, or a bearer token with one, is rejected rather than downgraded.
    #
    # 1. HMAC-signed: ApiKey carries the PUBLIC id (pk_live_…), and Signature +
    #    Timestamp must accompany it. The secret never crosses the wire.
    #
    #    X-Timestamp is unix SECONDS. The signed string is the timestamp, a
    #    literal ".", then the exact raw request body:
    #        X-Signature = hex( HMAC_SHA256( apiSecret, "${X-Timestamp}.${rawBody}" ) )
    #    For a request with no body (any GET) the raw body is "", so the signed
    #    string is "${X-Timestamp}.". Serialize the body once and send those
    #    exact bytes — re-serializing changes them and invalidates the
    #    signature. Requests more than 300 s from server time are rejected as
    #    replays. Verified against backend/src/middleware/auth.ts.
    ApiKey:      { type: apiKey, in: header, name: X-Api-Key }
    Signature:   { type: apiKey, in: header, name: X-Signature }
    Timestamp:   { type: apiKey, in: header, name: X-Timestamp }
    # 2. Bearer: the X-Api-Key value IS the secret (ak_live_…), sent alone.
    #    Simpler to integrate, but the credential is on the wire every call, so
    #    such keys are never granted the payouts:write scope.
    ApiKeyBearer: { type: apiKey, in: header, name: X-Api-Key }
    BearerAuth:  { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Invoice:
      type: object
      description: >
        A fiat-denominated document with line items. It mints no public token of
        its own: the customer-facing capability is `token`, which belongs to the
        invoice's single-use payment link.
      properties:
        id:       { type: string, example: inv_01HZY... }
        number:   { type: string, example: INV-0001, description: Sequential WITHIN this merchant }
        customerName:  { type: string, nullable: true }
        customerEmail: { type: string, nullable: true }
        currency: { type: string, example: INR }
        subtotal: { type: string, example: "3000.00" }
        taxAmount: { type: string, example: "450.00" }
        total:    { type: string, example: "3450.00" }
        notes:    { type: string, nullable: true }
        dueDate:  { type: string, format: date, nullable: true }
        status:   { type: string, enum: [draft, open, paid, void] }
        overdue:
          type: boolean
          description: >
            DERIVED from `dueDate` at read time, never stored — no job has to age
            a row into it, so an invoice cannot get stuck waiting for a clock.
        asset:    { type: string, nullable: true, description: Pinned settlement asset, or null to let the customer choose }
        network:  { type: string, nullable: true }
        token:
          type: string
          nullable: true
          description: >
            The pay link's token; the URL is {PUBLIC_PANEL_URL}/pay/{token}.
            Null while the invoice is a draft.
        paymentId: { type: string, nullable: true, description: The payment that settled it }
        subscriptionId: { type: string, nullable: true }
        cycleNumber:    { type: integer, nullable: true }
        issuedAt: { type: string, format: date-time, nullable: true }
        sentAt:   { type: string, format: date-time, nullable: true }
        paidAt:   { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          description: Empty in LIST responses; populated by GET /invoices/{id}.
          items: { $ref: '#/components/schemas/InvoiceItem' }
    InvoiceItem:
      type: object
      properties:
        id:          { type: string }
        description: { type: string }
        quantity:    { type: string, example: "2.5", description: May be fractional }
        unitPrice:   { type: string, example: "200" }
        amount:      { type: string, example: "500.00", description: quantity × unitPrice, stored }
    CreateInvoice:
      type: object
      required: [currency, items]
      properties:
        currency: { type: string, example: INR, description: Fiat code — see GET /rates }
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [description, unitPrice]
            properties:
              description: { type: string }
              quantity:    { type: string, default: "1" }
              unitPrice:   { type: string }
        customerName:  { type: string }
        customerEmail: { type: string, format: email }
        notes:    { type: string }
        dueDate:  { type: string, format: date, example: "2026-08-22" }
        taxAmount: { type: string }
        asset:    { type: string, description: Pin a settlement asset; requires `network` }
        network:  { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
        issue:
          type: boolean
          default: true
          description: Mint the pay link immediately. False saves a draft instead.
...
paths:
...
  /pay/{token}/invoice:
    get:
      tags: [Invoices]
      summary: "PUBLIC: the invoice behind a pay link"
      description: >
        An invoice's pay link IS its capability, so whoever holds the URL is
        exactly who should be able to read the document. Returns 404 when the
        link is an ordinary checkout link rather than an invoice — the checkout
        page probes this for every link and renders line items only if it gets
        something back.

        The response carries the document and the merchant's business name, and
        nothing else about the merchant: no id, email, wallet, balance, webhook,
        or other invoices.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  number:       { type: string }
                  merchantName: { type: string }
                  customerName: { type: string, nullable: true }
                  currency:     { type: string }
                  subtotal:     { type: string }
                  taxAmount:    { type: string }
                  total:        { type: string }
                  notes:        { type: string, nullable: true }
                  dueDate:      { type: string, format: date, nullable: true }
                  status:       { type: string, enum: [open, paid, void] }
                  overdue:      { type: boolean }
                  issuedAt:     { type: string, format: date-time, nullable: true }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        description: { type: string }
                        quantity:    { type: string }
                        unitPrice:   { type: string }
                        amount:      { type: string }
        "404": { description: This link has no invoice, or does not exist }
...
  /invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [draft, open, paid, void] } }
        - { in: query, name: limit, schema: { type: integer, default: 100, maximum: 200 } }
      responses:
        "200":
          description: OK. `items` is empty here — use GET /invoices/{id} for line items.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Invoice' } }
    post:
      tags: [Invoices]
      summary: Create an invoice
      description: >
        Invoicing is money-IN, so an API key of either mode may do it. Issuing
        mints a single-use payment link whose token is the customer's pay URL.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvoice' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        "400": { description: Bad request }
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: Retrieve an invoice, with line items
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        "404": { description: Not found }
  /invoices/{id}/send:
    post:
      tags: [Invoices]
      summary: Email the invoice to its customer
      description: >
        Rate limited per merchant per hour, on top of the usual key limiter:
        this is the only route that sends mail to an address the caller chooses,
        with content the caller wrote, which is a spam relay if left unbounded.

        `sent: false` means the gateway has no SMTP configured and logged the
        message instead — the documented development behaviour, not a failure.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  to:   { type: string }
        "400": { description: Still a draft, voided, or no customer email }
        "429": { description: Too many invoices sent this hour }
  /invoices/{id}/void:
    post:
      tags: [Invoices]
      summary: Withdraw an invoice and disable its pay link
      description: >
        Both happen together. An invoice marked void whose link still worked
        would be a withdrawn demand a customer could still pay. A PAID invoice
        cannot be voided — the money has moved.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        "400": { description: Already paid }
...
  /subscriptions/{id}/invoices:
    get:
      tags: [Subscriptions]
      summary: The invoices this plan has issued
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
...